no need to specify type

+++ Depo/Depo/Presentation/SubViews/PrivateShareSlider/SharedFilesCollectionDataSource.swift
@@ -64,10 +64,10 @@ extension SharedFilesCollectionDataSource: UICollectionViewDelegateFlowLayout {
     }
     
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return CGFloat(0)
+        return 0
     }
     
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return CGFloat(16)
+        return 16
     }
 }
